Revert python doc build to use mkdocs builder (#514)
This PR revert the python build for wherobots-compute to use mkdocs
because the downstream documents repository relies on the mkdocs other
than the Sphinx. We will need to investigate if Wherobots's python
document will switch to use the sphinx later, but this fix is to fix the
broken links downstream.