[BUG-1579] Fix "computeEnvelopeInternal is not implemented" when broadcast KNN join starts to spill (#536)
## Did you read the Contributor Guide?

- Yes, I have read the [Contributor
Rules](https://sedona.apache.org/latest/community/rule/) and
[Contributor Development
Guide](https://sedona.apache.org/latest/community/develop/)

## Is this PR related to a ticket?

- Yes. The notion ticket is BUG-1579.

## What changes were proposed in this PR?

The query side of the KNN join is wrapped as a `UniqueGeometry` when the
query side is broadcasted. When the spatial index was spilled, the query
side will be sorted by the hilbert index of their bounding boxes.
`computeEnvelopeInternal` will be called to obtain the bounding boxes of
query side geometries. Unfortunately, `UniqueGeometry` implements
`computeEnvelopeInternal` by throwing a NotImplementedException, this
causes the problem. This PR implements the `computeEnvelopeInternal`
method of `UniqueGeometry` by delegating it to the original geometry.

## How was this patch tested?

Passing newly added tests

## Did this PR include necessary documentation updates?

- No, this PR does not affect any public API so no need to change the
documentation.