[Perf] Improve perf of pixelAsPolygons (#540)
A couple of small changes to improve the performance of
`pixelAsPolygons`.

## Changes

Profiling showed a few opportunities.

1. found that the majority of time was spent in `createPolygon`, but
creating the same number of polygons in `polygonize` did not show this
is the performance analysis, so copied the way that it creates the
polygons.

2. saw a lot of time was spent in a bunch of inits. Moved up a bunch
outside of the loop.

3. `PackedCoordinateSequenceFactory` to store coordinate data in a
packed array, optimizing for memory efficiency and pre-allocate space
for the PixelRecords.

## Task


https://www.notion.so/wherobots/Investigate-performance-issue-of-RS_PixelAsPolygons-2a9ba32ec02780c9be3be5321e584484

## Performance Testing

With these two changes, the time to create 1 million polygons 100 times
dropped from 14.8 seconds to 6.8 seconds.

## Testing

All existing unit tests pass.